+++ crates/services/executor/src/ports.rs
@@ -164,7 +164,7 @@ pub trait TransactionsSource {
         &self,
         gas_limit: u64,
         tx_count_limit: u32,
-        block_transaction_size_limit: u32,
+        block_transaction_size_limit: u64,
     ) -> Vec<MaybeCheckedTransaction>;
 }
